requirements.txt
asgiref==3.6.0
Django==4.2.1
django-debug-toolbar==4.0.0
Pillow==9.5.0
sqlparse==0.4.4
